Gate slow retry test behind integration build tag
- Move 6-second retry test to separate file with +build integration
- Keeps make test fast for normal development
- Run with -tags=integration when needed
